Moved to Single Feature Service Layer
Moving to Single Service Layer.
Remnants of the old 4 Feature Service Layer model are still present. and
there are a number of hacks that are in place to make objects that will
be built with functions.
Will need to add the code to allow features to be added and removed.